+++ server/src/routes/guests.js
@@ -5,6 +5,7 @@ import {
   getGuest,
   createGuest,
   updateGuest,
+  deleteGuest,
   checkoutGuest,
   renewGuestStay,
   getGuestCommunications,
@@ -23,6 +24,7 @@ router.get('/stats', getGuestStats)
 router.post('/', createGuest)
 router.get('/:id', getGuest)
 router.put('/:id', updateGuest)
+router.delete('/:id', deleteGuest)
 router.post('/:id/documents', uploadDocs.array('documents', 12), uploadGuestDocuments)
 router.post('/:id/checkout', checkoutGuest)
 router.post('/:id/renew', renewGuestStay)